\name{JMbayesCpp-package}
\alias{JMbayesCpp-package}
\alias{JMbayesCpp}
\docType{package}
\title{
\packageTitle{JMbayesCpp}
}
\description{
\packageDescription{JMbayesCpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{JMbayesCpp}
\packageIndices{JMbayesCpp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{JMbayesCpp}

Maintainer: \packageMaintainer{JMbayesCpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
